refactor: rename CreatorReviewCard and move to shared components
